move code, fixups
